requirements.txt
Django==5.0.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.3.1
psycopg2-binary==2.9.9
python-dotenv==1.0.1
drf-yasg==1.21.7
django-filter==23.5
msgpack==1.0.7
django-redis==5.4.0
django-guardian==2.4.0 